adding 2nd db for testing which has consistent data

+++ src/main/java/com/napier/sem/Continent.java
@@ -5,11 +5,11 @@ public class Continent {
     // Continent name
     public String continent_name;
     // Continent population
-    public int continent_population;
+    public Long continent_population;
 
     // Continent city population
-    public int continent_city_population;
+    public Long continent_city_population;
 
     // Continent non-city population
-    public int continent_non_city_population;
+    public Long continent_non_city_population;
 }
